Refactor Tracing Utils
