+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/pages/tracing/utils/span.js
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-import Constants from "./Constants";
+import Constants from "./constants";
 
 /**
  * Single span in a Trace.
